+++ integreat_cms/cms/views/pages/page_actions.py
@@ -16,6 +16,7 @@ from django.utils.translation import ugettext as _
 from django.views.decorators.http import require_POST
 from django.db import transaction
 
+from linkcheck.models import Link
 from treebeard.exceptions import InvalidPosition, InvalidMoveToDescendant
 
 from ....api.decorators import json_response
@@ -58,6 +59,10 @@ def archive_page(request, page_id, region_slug, language_slug):
         raise PermissionDenied(
             f"{request.user!r} does not have the permission to archive {page!r}"
         )
+
+    # Delete related link objects as they are no longer required
+    for child_page in Page.get_tree(parent=page).exclude(explicitly_archived=True):
+        Link.objects.filter(page_translation__page=child_page).delete()
 
     page.explicitly_archived = True
     page.save()
@@ -129,6 +134,13 @@ def restore_page(request, page_id, region_slug, language_slug):
                 "language_slug": language_slug,
             },
         )
+
+    # Restore related link objects
+    for child_page in Page.get_tree(parent=page).cache_tree(archived=False):
+        for translation in child_page.translations.distinct("page__pk", "language__pk"):
+            # The post_save signal will create link objects from the content
+            translation.save(update_timestamp=False)
+
     logger.debug("%r restored by %r", page, request.user)
     messages.success(request, _("Page was successfully restored."))
     return redirect(